Removed the unused NexusAttribute.Versioning property.
