+++ src/NexNet.Generator/MetaGenerator/NexusAttributeMeta.cs
@@ -6,8 +6,6 @@ internal class NexusAttributeMeta : AttributeMetaBase
 {
     public bool IsClient { get; private set; }
     public bool IsServer { get; private set; }
-    public bool VersionMustMatch { get; private set; } = true;
-    public bool VersionNegotiation { get; private set; }
 
     public NexusAttributeMeta(INamedTypeSymbol symbol)
         : base("NexusAttribute", symbol)
@@ -20,11 +18,6 @@ internal class NexusAttributeMeta : AttributeMetaBase
         {
             IsClient = (int)GetItem(typedConstant)! == 0;
             IsServer = (int)GetItem(typedConstant)! == 1;
-        }
-        else if (key == "Versioning" || constructorArgIndex == 1)
-        {
-            VersionMustMatch = (int)GetItem(typedConstant)! == 0;
-            VersionNegotiation = (int)GetItem(typedConstant)! == 1;
         }
     }
 }
